elx-package-metadata: Avoid dynamic-scope-related naming conflict.
Using the variable `features' is problematic since Emacs will joyfully update
that value when loading new libraries. If a new feature is `provide'd in the
middle of `elx-package-metadata', such as when `mapcan' is encountered as a
result of an `autoload', the value of `features' is updated with the new
features provided. This causes spurious features to appear in the `features'
variable used by `elx-package-metadata'.

This was one hell of an annoying heisenbug to find.

Signed-off-by: Dan Hackney <dan@haxney.org>